Refactor get_filename_from_url
