+++ saleor/graphql/core/validators/file.py
@@ -1,13 +1,12 @@
 import mimetypes
 import os
-import secrets
 
 import requests
 from django.core.exceptions import ValidationError
 from PIL import Image, UnidentifiedImageError
 
 from ....thumbnail import MIME_TYPE_TO_PIL_IDENTIFIER
-from ..utils import FILE_NAME_MAX_LENGTH, add_hash_to_file_name
+from ..utils import add_hash_to_file_name
 
 Image.init()
 
@@ -48,15 +47,6 @@ def validate_image_url(url: str, field_name: str, error_code: str) -> None:
         raise ValidationError(
             {field_name: ValidationError("Invalid file type.", code=error_code)}
         )
-
-
-def get_filename_from_url(url: str) -> str:
-    """Prepare unique filename for file from URL to avoid overwritting."""
-    file_name = os.path.basename(url)
-    name, format = os.path.splitext(file_name)
-    name = name[:FILE_NAME_MAX_LENGTH]
-    hash = secrets.token_hex(nbytes=4)
-    return f"{name}_{hash}{format}"
 
 
 def clean_image_file(cleaned_input, img_field_name, error_class):
