+++ saleor/thumbnail/tests/test_utils.py
@@ -4,10 +4,11 @@ import graphene
 import pytest
 from django.core.files import File
 
-from .. import ThumbnailFormat
+from .. import FILE_NAME_MAX_LENGTH, ThumbnailFormat
 from ..models import Thumbnail
 from ..utils import (
     ProcessedImage,
+    get_filename_from_url,
     get_image_or_proxy_url,
     get_thumbnail_size,
     prepare_image_proxy_url,
@@ -112,3 +113,52 @@ def test_processed_image_preprocess_method_called(category_with_image, thumb_for
 
     # then
     preprocess_mock.assert_called_once()
+
+
+def test_get_filename_from_url_unique():
+    # given
+    file_format = "jpg"
+    file_name = "lenna"
+    url = f"http://example.com/{file_name}.{file_format}"
+
+    # when
+    result = get_filename_from_url(url)
+
+    # then
+    assert result.startswith(file_name)
+    assert result.endswith(file_format)
+    assert result != f"{file_name}.{file_format}"
+
+
+def test_get_filename_from_url_with_long_name_is_trimmed():
+    # given
+    file_format = "jpg"
+    file_name = "2Fvar2Ffolders2Fbj2F61gtb14j7rz474yd15tnkzjh0000gn2FT2Fa"
+    assert len(file_name) > FILE_NAME_MAX_LENGTH
+    url = f"http://example.com/{file_name}.{file_format}"
+
+    # when
+    result = get_filename_from_url(url)
+
+    # then
+    assert result.startswith(file_name[:FILE_NAME_MAX_LENGTH])
+    assert result.endswith(file_format)
+    assert result != f"{file_name}.{file_format}"
+    assert len(result.split("_")[0]) == FILE_NAME_MAX_LENGTH
+
+
+def test_get_filename_from_url_with_short_name_is_not_trimmed():
+    # given
+    file_format = "jpg"
+    file_name = "short"
+    assert len(file_name) < FILE_NAME_MAX_LENGTH
+    url = f"http://example.com/{file_name}.{file_format}"
+
+    # when
+    result = get_filename_from_url(url)
+
+    # then
+    assert result.startswith(file_name)
+    assert result.endswith(file_format)
+    assert result != f"{file_name}.{file_format}"
+    assert len(result.split("_")[0]) < FILE_NAME_MAX_LENGTH
